Give the aircraft own return its own test module
The traffic tests crossed the file size limit, and the tests that cross it prove
the opposite of what the file is named for: that the aircraft own return is a
position and not a target. They read better beside that claim.
