+++ crates/pilotage-presentation/src/tests.rs
@@ -2,6 +2,7 @@
 
 mod advisory;
 mod layer;
+mod ownship;
 mod traffic;
 mod vertical;
 mod weather;
